Move path.go into taskset directory
