+++ sources/mdtk/taskset/read/read_taskfile_test.go
@@ -1,7 +1,7 @@
 package read
 
 import (
-	"mdtk/path"
+	"mdtk/taskset/path"
 	"testing"
 	"github.com/stretchr/testify/assert"
 )
